Experiment: single-word phase to QueryEvaluator
Tried to add an additional phase to QueryEvaluator, where it scans for
single-word matches in the wordlist before starting to build up the
datastructures for the PhraseMatcher.

This would theoretically improve response times for complex expressions
(with many states) on long wordlists, where the result is a single word.

In practice, it increased the runtime of most searches:

- If there's a single-word match, it's "`O(n)`-ish" to find it anyway

- Even with complex expressions with many states, usually states can get
  pruned pretty quickly (by being nilpotent)

- First pass is complete "throw-away" work for the later steps; usually
  we want to at least *check* for phrase matches, even if there were
  single-word matches.

The refactor to add `is_word_match` seemed to also cause a perf hit?
It would have been nice to leave to round out the API...