WIP: Use `fst` crate for wordlist; change match iter order
Shuffle the order of iteration between query/matcher/expression.
Previously the single-word matching iterated in:
    - word in vec, expression, char in word, ...
Instead, do the following:
    - word in fst by char, expression, ...
This avoids needing to re-compute the shared prefix between words, and
makes it easier to iterate over a trie-like wordlist (i.e. an
`fst::Map`).

This has the side effect of making the order of expressions less
important, because they can be evaluated one character at a time (with
early termination). This makes `<pzszta> !3; [piza]{5}` now run in about
the same time as `[piza]{5}; <pzszta> !3`.

So far, it's looking like about a ~4x speed improvement on my laptop.
(Some of those gains will likely be re-lost when doing word lookup?)

If it's sufficiently fast, we can get away with the downside that the
single-pass matches will come out in alphabetical order (instead of by
tranche).